update: change IDs and triggeredForCatalogItem to long, bump to 11.0.1

+++ src/main/java/one/talon/model/Account.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Account {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -150,7 +150,7 @@ public class Account {
   private Object attributes;
 
 
-  public Account id(Integer id) {
+  public Account id(Long id) {
     
     this.id = id;
     return this;
@@ -162,12 +162,12 @@ public class Account {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
