+++ src/main/java/one/talon/model/AccountAdditionalCost.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class AccountAdditionalCost {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -114,7 +114,7 @@ public class AccountAdditionalCost {
   private TypeEnum type = TypeEnum.SESSION;
 
 
-  public AccountAdditionalCost id(Integer id) {
+  public AccountAdditionalCost id(Long id) {
     
     this.id = id;
     return this;
@@ -126,12 +126,12 @@ public class AccountAdditionalCost {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
