+++ src/main/java/one/talon/model/Achievement.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Achievement {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -246,7 +246,7 @@ public class Achievement {
   private StatusEnum status;
 
 
-  public Achievement id(Integer id) {
+  public Achievement id(Long id) {
     
     this.id = id;
     return this;
@@ -258,12 +258,12 @@ public class Achievement {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
