+++ src/main/java/one/talon/model/Attribute.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Attribute {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -272,7 +272,7 @@ public class Attribute {
   private Integer eventTypeId;
 
 
-  public Attribute id(Integer id) {
+  public Attribute id(Long id) {
     
     this.id = id;
     return this;
@@ -284,12 +284,12 @@ public class Attribute {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
