+++ src/main/java/one/talon/model/Audience.java
@@ -36,7 +36,7 @@ public class Audience {
 
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -93,7 +93,7 @@ public class Audience {
   }
 
 
-  public Audience id(Integer id) {
+  public Audience id(Long id) {
     
     this.id = id;
     return this;
@@ -105,12 +105,12 @@ public class Audience {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
