+++ src/main/java/one/talon/model/CampaignCollectionWithoutPayload.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CampaignCollectionWithoutPayload {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -71,7 +71,7 @@ public class CampaignCollectionWithoutPayload {
   private Integer campaignId;
 
 
-  public CampaignCollectionWithoutPayload id(Integer id) {
+  public CampaignCollectionWithoutPayload id(Long id) {
     
     this.id = id;
     return this;
@@ -83,12 +83,12 @@ public class CampaignCollectionWithoutPayload {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
