+++ src/main/java/one/talon/model/CampaignGroup.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CampaignGroup {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -65,7 +65,7 @@ public class CampaignGroup {
   private List<Integer> campaignIds = null;
 
 
-  public CampaignGroup id(Integer id) {
+  public CampaignGroup id(Long id) {
     
     this.id = id;
     return this;
@@ -77,12 +77,12 @@ public class CampaignGroup {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
