+++ src/main/java/one/talon/model/CatalogItem.java
@@ -37,7 +37,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CatalogItem {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -68,7 +68,7 @@ public class CatalogItem {
   private Product product;
 
 
-  public CatalogItem id(Integer id) {
+  public CatalogItem id(Long id) {
     
     this.id = id;
     return this;
@@ -80,12 +80,12 @@ public class CatalogItem {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
