+++ src/main/java/one/talon/model/Collection.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Collection {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -81,7 +81,7 @@ public class Collection {
   private List<String> payload = null;
 
 
-  public Collection id(Integer id) {
+  public Collection id(Long id) {
     
     this.id = id;
     return this;
@@ -93,12 +93,12 @@ public class Collection {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
