+++ src/main/java/one/talon/model/CollectionWithoutPayload.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CollectionWithoutPayload {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -77,7 +77,7 @@ public class CollectionWithoutPayload {
   private Integer campaignId;
 
 
-  public CollectionWithoutPayload id(Integer id) {
+  public CollectionWithoutPayload id(Long id) {
     
     this.id = id;
     return this;
@@ -89,12 +89,12 @@ public class CollectionWithoutPayload {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
