+++ src/main/java/one/talon/model/CouponCreationJob.java
@@ -36,7 +36,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CouponCreationJob {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -123,7 +123,7 @@ public class CouponCreationJob {
   private Integer chunkSize;
 
 
-  public CouponCreationJob id(Integer id) {
+  public CouponCreationJob id(Long id) {
     
     this.id = id;
     return this;
@@ -135,12 +135,12 @@ public class CouponCreationJob {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
