+++ src/main/java/one/talon/model/CouponDeletionJob.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CouponDeletionJob {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -82,7 +82,7 @@ public class CouponDeletionJob {
   private List<Integer> campaignIDs = null;
 
 
-  public CouponDeletionJob id(Integer id) {
+  public CouponDeletionJob id(Long id) {
     
     this.id = id;
     return this;
@@ -94,12 +94,12 @@ public class CouponDeletionJob {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
