+++ src/main/java/one/talon/model/CustomerProfile.java
@@ -37,7 +37,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CustomerProfile {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -80,7 +80,7 @@ public class CustomerProfile {
   private Boolean sandbox;
 
 
-  public CustomerProfile id(Integer id) {
+  public CustomerProfile id(Long id) {
     
     this.id = id;
     return this;
@@ -92,12 +92,12 @@ public class CustomerProfile {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
