+++ src/main/java/one/talon/model/Environment.java
@@ -45,7 +45,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Environment {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -104,7 +104,7 @@ public class Environment {
   private List<ApplicationCIF> applicationCartItemFilters = null;
 
 
-  public Environment id(Integer id) {
+  public Environment id(Long id) {
     
     this.id = id;
     return this;
@@ -116,12 +116,12 @@ public class Environment {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
