+++ src/main/java/one/talon/model/Event.java
@@ -36,7 +36,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Event {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -79,7 +79,7 @@ public class Event {
   private Meta meta;
 
 
-  public Event id(Integer id) {
+  public Event id(Long id) {
     
     this.id = id;
     return this;
@@ -91,12 +91,12 @@ public class Event {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
