+++ src/main/java/one/talon/model/Giveaway.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Giveaway {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -75,7 +75,7 @@ public class Giveaway {
   private Integer profileId;
 
 
-  public Giveaway id(Integer id) {
+  public Giveaway id(Long id) {
     
     this.id = id;
     return this;
@@ -87,12 +87,12 @@ public class Giveaway {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
