+++ src/main/java/one/talon/model/InventoryCoupon.java
@@ -36,7 +36,7 @@ import org.threeten.bp.OffsetDateTime;
 public class InventoryCoupon {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -131,7 +131,7 @@ public class InventoryCoupon {
   private String state;
 
 
-  public InventoryCoupon id(Integer id) {
+  public InventoryCoupon id(Long id) {
     
     this.id = id;
     return this;
@@ -143,12 +143,12 @@ public class InventoryCoupon {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
