+++ src/main/java/one/talon/model/InventoryReferral.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class InventoryReferral {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -89,7 +89,7 @@ public class InventoryReferral {
   private List<String> referredCustomers = new ArrayList<String>();
 
 
-  public InventoryReferral id(Integer id) {
+  public InventoryReferral id(Long id) {
     
     this.id = id;
     return this;
@@ -101,12 +101,12 @@ public class InventoryReferral {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
