+++ src/main/java/one/talon/model/LoyaltyCard.java
@@ -38,7 +38,7 @@ import org.threeten.bp.OffsetDateTime;
 public class LoyaltyCard {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -101,7 +101,7 @@ public class LoyaltyCard {
   private String batchId;
 
 
-  public LoyaltyCard id(Integer id) {
+  public LoyaltyCard id(Long id) {
     
     this.id = id;
     return this;
@@ -113,12 +113,12 @@ public class LoyaltyCard {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
