+++ src/main/java/one/talon/model/LoyaltyProgram.java
@@ -37,7 +37,7 @@ import org.threeten.bp.OffsetDateTime;
 public class LoyaltyProgram {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -340,7 +340,7 @@ public class LoyaltyProgram {
   private Boolean canUpdateSubledgers = false;
 
 
-  public LoyaltyProgram id(Integer id) {
+  public LoyaltyProgram id(Long id) {
     
     this.id = id;
     return this;
@@ -352,12 +352,12 @@ public class LoyaltyProgram {
   **/
   @ApiModelProperty(example = "139", required = true, value = "The ID of loyalty program.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
