+++ src/main/java/one/talon/model/LoyaltyTier.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class LoyaltyTier {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -61,7 +61,7 @@ public class LoyaltyTier {
   private BigDecimal minPoints;
 
 
-  public LoyaltyTier id(Integer id) {
+  public LoyaltyTier id(Long id) {
     
     this.id = id;
     return this;
@@ -73,12 +73,12 @@ public class LoyaltyTier {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
