+++ src/main/java/one/talon/model/MultipleAudiencesItem.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class MultipleAudiencesItem {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -100,7 +100,7 @@ public class MultipleAudiencesItem {
   private StatusEnum status;
 
 
-  public MultipleAudiencesItem id(Integer id) {
+  public MultipleAudiencesItem id(Long id) {
     
     this.id = id;
     return this;
@@ -112,12 +112,12 @@ public class MultipleAudiencesItem {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
