+++ src/main/java/one/talon/model/Picklist.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Picklist {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -116,7 +116,7 @@ public class Picklist {
   private Boolean imported;
 
 
-  public Picklist id(Integer id) {
+  public Picklist id(Long id) {
     
     this.id = id;
     return this;
@@ -128,12 +128,12 @@ public class Picklist {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
