+++ src/main/java/one/talon/model/Referral.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Referral {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -83,7 +83,7 @@ public class Referral {
   private String batchId;
 
 
-  public Referral id(Integer id) {
+  public Referral id(Long id) {
     
     this.id = id;
     return this;
@@ -95,12 +95,12 @@ public class Referral {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
