+++ src/main/java/one/talon/model/Role.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class Role {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -69,7 +69,7 @@ public class Role {
   private Object acl;
 
 
-  public Role id(Integer id) {
+  public Role id(Long id) {
     
     this.id = id;
     return this;
@@ -81,12 +81,12 @@ public class Role {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
