+++ src/main/java/one/talon/model/RoleV2.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class RoleV2 {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -66,7 +66,7 @@ public class RoleV2 {
   private List<Integer> members = null;
 
 
-  public RoleV2 id(Integer id) {
+  public RoleV2 id(Long id) {
     
     this.id = id;
     return this;
@@ -78,12 +78,12 @@ public class RoleV2 {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
