+++ src/main/java/one/talon/model/WebhookWithOutgoingIntegrationDetails.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class WebhookWithOutgoingIntegrationDetails {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -147,7 +147,7 @@ public class WebhookWithOutgoingIntegrationDetails {
   private String outgoingIntegrationTypeName;
 
 
-  public WebhookWithOutgoingIntegrationDetails id(Integer id) {
+  public WebhookWithOutgoingIntegrationDetails id(Long id) {
     
     this.id = id;
     return this;
@@ -159,12 +159,12 @@ public class WebhookWithOutgoingIntegrationDetails {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
